Conflicts:
	www/css/base.css

+++ www/core/modules/data/ajax/file.ajax.php
@@ -281,7 +281,7 @@ class FileAjax
 					$permission = DataAjax::permission_window();
 					$button_handler_template = new JSTemplate("data/js/file_permission_window.js");
 					$button_handler_template->set_var("session_id", $_GET['session_id']);
-					$button_handler_template->set_var("file_id", $_GET['file_id']);
+					$button_handler_template->set_var("file_id", $_POST['file_id']);
 					$button_handler = $button_handler_template->get_string();
 					$button_handler_caption = "Change";
 					$html_caption = "Change permission";
@@ -298,7 +298,7 @@ class FileAjax
 					$template = new HTMLTemplate("data/file_delete_window.html");
 					$button_handler_template = new JSTemplate("data/js/file_delete_window.js");
 					$button_handler_template->set_var("session_id", $_GET['session_id']);
-					$button_handler_template->set_var("file_id", $_GET['file_id']);
+					$button_handler_template->set_var("file_id", $_POST['file_id']);
 					$button_handler = $button_handler_template->get_string();
 					$button_handler_caption = "Delete";
 					$html_caption = "Delete File";
@@ -312,31 +312,48 @@ class FileAjax
 	
 	public static function add_file($folder_id)
 	{
-		$paramquery = $_GET;
-		$unique_id = uniqid();
-		$paramquery[unique_id] = $unique_id;
-		$params = http_build_query($paramquery);
-		$template = new HTMLTemplate("data/file_upload_window.html");
-		$template->set_var("params", $params);
-		$template->set_var("unique_id", $unique_id);
-		$template->set_var("session_id", $_GET[session_id]);
-		$button_handler_template = new JSTemplate("data/js/file_upload_window.js");
-		$button_handler = $button_handler_template->get_string();
-		$button_handler_caption = "Add";
-		$html_caption = "Add File";
-		$html = $template->get_string();
-		$additional_script_template = new JSTemplate("data/js/file_upload_window_additional.js");
-		$additional_script_template->set_var("session_id", $_GET['session_id']);
-		$additional_script_template->set_var("unique_id", $unique_id);
-		$additional_script = $additional_script_template->get_string();
-		$array = array("content"=>$html , "content_caption"=>$html_caption , "handler"=>$button_handler , "handler_caption"=>$button_handler_caption, "additional_script"=>$additional_script);
-		return json_encode($array);
+		$parent_folder = Folder::get_instance($folder_id);
+		if ($parent_folder->is_write_access())
+		{
+			$paramquery = array();
+			$unique_id = uniqid();
+			$paramquery[session_id] = $_GET[session_id];
+			$paramquery[folder_id] = $folder_id;
+			$paramquery[unique_id] = $unique_id;
+			$params = http_build_query($paramquery);
+			$template = new HTMLTemplate("data/file_upload_window.html");
+			$template->set_var("params", $params);
+			$template->set_var("unique_id", $unique_id);
+			$template->set_var("session_id", $_GET[session_id]);
+			$button_handler_template = new JSTemplate("data/js/file_upload_window.js");
+			$button_handler = $button_handler_template->get_string();
+			$button_handler_caption = "Add";
+			$html_caption = "Add File";
+			$html = $template->get_string();
+			$additional_script_template = new JSTemplate("data/js/file_upload_window_additional.js");
+			$additional_script_template->set_var("session_id", $_GET['session_id']);
+			$additional_script_template->set_var("unique_id", $unique_id);
+			$additional_script = $additional_script_template->get_string();
+			$array = array("content"=>$html , "content_caption"=>$html_caption , "handler"=>$button_handler , "handler_caption"=>$button_handler_caption, "additional_script"=>$additional_script);
+			return json_encode($array);
+		}
+		else
+		{
+			throw new DataSecurityAccessDeniedException();
+		}
 	}
 	
 	private static function delete_file($file_id) 
 	{
 		$file = File::get_instance($file_id);
-		$file->delete();
+		if ($file->is_delete_access())
+		{
+			$file->delete();
+		}
+		else
+		{
+			throw new DataSecurityAccessDeniedException();
+		}
 	}
 }
